+++ authority-portal-frontend/src/app/core/utils/validators/named-regex-validator.ts
@@ -1,0 +1,10 @@
+import {AbstractControl, ValidationErrors, ValidatorFn} from '@angular/forms';
+
+export const namedRegexValidator =
+  (regex: RegExp, key: string): ValidatorFn =>
+  (c: AbstractControl): ValidationErrors | null => {
+    if (!c.value) {
+      return null;
+    }
+    return regex.test(c.value) ? null : {[key]: true};
+  };